Simplify Rack::Protection::SessionHijack
The underlying sessions are now fully encrypted. We can just rely on
this feature to protect the header canaries used by SessionHijack.

Additionally, This middleware is now more useful. A theoretical attack
is now unable to view the session data and would not know what to forge
the stored User-Agent and Language headers to.